Add encoder to batch insert interface
The previous batch insert interface allowed the user to specify
invalid JSON which lead to confusing errors and strange interpolation.
The new interface requires the user to provide a proper encoder
```ocaml
val encoder : 'a -> Js.Json.t array
```
to encode all of the values given in the `~rows` parameter.  This also
means that `~rows` can now be an array of any `'a` ocaml value because
the `~encoder` is responsible for the proper transformation of those
values.